requirements.txt
mysqlclient
django
djangorestframework
gunicorn